price-update-agent/pyproject.toml
[tool.poetry]
name = "price-update-agent"
version = "0.1.0"
description = "Specialized agent for executing menu price updates, item modifications, and availability changes through natural language commands with database write operations"
authors = ["Club Management System <dev@clubmanagement.com>"]
license = "Apache-2.0"
readme = "README.md"
packages = [{include = "price_update_agent"}]

[tool.poetry.dependencies]
python = "^3.13"
google-adk = { version = "^1.2.0", extras = ["mcp"] }
pyyaml = "^6.0"
fastapi = "^0.115.0"
uvicorn = "^0.34.0"
google-cloud-aiplatform = { extras = ["adk", "agent-engines"], version = "^1.93.0" }
pydantic = "^2.10.6"
python-dotenv = "^1.0.1"
google-genai = "^1.16.1"
httpx = "^0.28.0"
aiohttp = "^3.10.0"
sse-starlette = "^2.3.0"
pyjwt = "^2.10.0"
jwcrypto = "^1.5.0"
a2a-sdk = "^0.2.12"
litellm = "^1.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
pytest-asyncio = "^0.26.0"
psutil = "^5.9.0"
requests = "^2.31.0"

[tool.poetry.group.deployment]
optional = true

[tool.poetry.group.deployment.dependencies]
absl-py = "^2.2.1"
cloudpickle = "^3.1.1"
flake8-pyproject = "^1.2.3"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "price-update-agent"
version = "0.1.0"
dependencies = [
    "a2a-sdk>=0.2.12",
    "fastapi>=0.115.0",
    "uvicorn>=0.34.0",
    "httpx>=0.28.0",
    "pydantic>=2.10.0",
    "python-dotenv>=1.0.1",
]

[project.scripts]
start = "price_update_agent.__main__:main" 